stateless processor start
